Update to spice-gtk 0.27

+++ jni/src/spice-common/common/log.c
@@ -84,7 +84,7 @@ void spice_logv(const char *log_domain,
         abort_level = getenv("SPICE_ABORT_LEVEL") ? atoi(getenv("SPICE_ABORT_LEVEL")) : SPICE_ABORT_LEVEL_DEFAULT;
     }
 
-    if (debug_level < log_level)
+    if (debug_level < (int) log_level)
         return;
 
     fprintf(stderr, "(%s:%d): ", getenv("_"), getpid());
@@ -104,7 +104,7 @@ void spice_logv(const char *log_domain,
 
     fprintf(stderr, "\n");
 
-    if (abort_level != -1 && abort_level >= log_level) {
+    if (abort_level != -1 && abort_level >= (int) log_level) {
         spice_backtrace();
         abort();
     }
